backend/requirements.dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.dev.txt requirements.dev.in
#
aniso8601==7.0.0
    # via
    #   -r requirements.txt
    #   graphene
asgiref==3.3.4
    # via
    #   -r requirements.txt
    #   django
attrs==20.3.0
    # via pytest
backcall==0.2.0
    # via ipython
bleach==3.3.0
    # via
    #   -r requirements.txt
    #   django-todo
cli-helpers[styles]==2.1.0
    # via pgcli
click==7.1.2
    # via
    #   pgcli
    #   pgspecial
configobj==5.0.6
    # via
    #   cli-helpers
    #   pgcli
decorator==5.0.7
    # via ipython
django-autocomplete-light==3.8.2
    # via
    #   -r requirements.txt
    #   django-todo
django-cors-headers==3.7.0
    # via -r requirements.txt
django-environ==0.4.5
    # via -r requirements.txt
django-extensions==3.1.2
    # via -r requirements.dev.in
django-filter==2.4.0
    # via
    #   -r requirements.txt
    #   django-graphql-auth
django-graphql-auth==0.3.16
    # via -r requirements.txt
django-graphql-jwt==0.3.0
    # via
    #   -r requirements.txt
    #   django-graphql-auth
git+git://github.com/jnoortheen/django-pgcli@master#egg=django-pgcli
    # via -r requirements.dev.in
git+git://github.com/mnieber/django-todo@master#egg=django-todo
    # via -r requirements.txt
django==3.2
    # via
    #   -r requirements.txt
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-graphql-auth
    #   django-graphql-jwt
    #   django-pgcli
    #   django-todo
    #   graphene-django
factory-boy==3.2.0
    # via -r requirements.dev.in
faker==8.1.0
    # via factory-boy
graphene-django==2.15.0
    # via
    #   -r requirements.txt
    #   django-graphql-auth
    #   django-graphql-jwt
graphene==2.1.8
    # via
    #   -r requirements.txt
    #   django-graphql-auth
    #   graphene-django
graphql-core==2.3.2
    # via
    #   -r requirements.txt
    #   django-graphql-jwt
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via
    #   -r requirements.txt
    #   graphene
iniconfig==1.1.1
    # via pytest
ipython-genutils==0.2.0
    # via traitlets
ipython==7.22.0
    # via -r requirements.dev.in
jedi==0.18.0
    # via ipython
packaging==20.9
    # via
    #   -r requirements.txt
    #   bleach
    #   pytest
parso==0.8.2
    # via jedi
pendulum==2.1.2
    # via pgcli
pexpect==4.8.0
    # via ipython
pgcli==3.1.0
    # via django-pgcli
pgspecial==1.12.1
    # via pgcli
pickleshare==0.7.5
    # via ipython
pluggy==0.13.1
    # via pytest
promise==2.3
    # via
    #   -r requirements.txt
    #   graphene-django
    #   graphql-core
    #   graphql-relay
prompt-toolkit==3.0.18
    # via
    #   ipython
    #   pgcli
psycopg2==2.8.6
    # via
    #   -r requirements.txt
    #   pgcli
    #   pgspecial
ptyprocess==0.7.0
    # via pexpect
pudb==2020.1
    # via -r requirements.dev.in
py==1.10.0
    # via pytest
pygments==2.8.1
    # via
    #   cli-helpers
    #   ipython
    #   pgcli
    #   pudb
pyjwt==1.7.1
    # via
    #   -r requirements.txt
    #   django-graphql-auth
    #   django-graphql-jwt
pyparsing==2.4.7
    # via
    #   -r requirements.txt
    #   packaging
pytest-html==3.1.1
    # via -r requirements.dev.in
pytest-metadata==1.11.0
    # via pytest-html
pytest==6.2.3
    # via
    #   -r requirements.dev.in
    #   pytest-html
    #   pytest-metadata
python-dateutil==2.8.1
    # via
    #   faker
    #   pendulum
pytz==2021.1
    # via
    #   -r requirements.txt
    #   django
pytzdata==2020.1
    # via pendulum
rx==1.6.1
    # via
    #   -r requirements.txt
    #   graphql-core
setproctitle==1.2.2
    # via pgcli
singledispatch==3.6.1
    # via
    #   -r requirements.txt
    #   graphene-django
six==1.15.0
    # via
    #   -r requirements.txt
    #   bleach
    #   configobj
    #   django-autocomplete-light
    #   graphene
    #   graphene-django
    #   graphql-core
    #   graphql-relay
    #   promise
    #   python-dateutil
    #   singledispatch
sqlparse==0.4.1
    # via
    #   -r requirements.txt
    #   django
    #   pgcli
    #   pgspecial
tabulate[widechars]==0.8.9
    # via cli-helpers
terminaltables==3.1.0
    # via cli-helpers
text-unidecode==1.3
    # via
    #   -r requirements.txt
    #   faker
    #   graphene-django
toml==0.10.2
    # via pytest
traitlets==5.0.5
    # via ipython
unidecode==1.2.0
    # via
    #   -r requirements.txt
    #   django-todo
urwid==2.1.2
    # via pudb
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   -r requirements.txt
    #   bleach

# The following packages are considered to be unsafe in a requirements file:
# setuptools
